services/ingestion/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
kafka-python
pydantic
pydantic-settings
structlog
python-json-logger
prometheus-client==0.19.0
minio
